requirements.txt
fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0
SQLAlchemy==2.0.42
requests==2.32.4
python-multipart==0.0.20
bcrypt==4.3.0
python-jose==3.5.0
pydantic[email]
pydantic-settings==2.10.1
coverage==7.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-asyncio==1.1.0
pytest-mock==3.14.1
httpx==0.28.1
cryptography==45.0.5
python-dotenv==1.1.1
pymysql==1.1.1
flake8==7.3.0
Pillow==10.4.0
PyJWT==2.10.1
passlib==1.7.4